// code/frontend/apps/game-index/src/gen/catalogApi/types/GamesSearch.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type GamesSearch = {
    /**
     * @type null,integer, int64
     */
    id?: number | null;
    /**
     * @type null,string
     */
    name?: string | null;
    /**
     * @type null,string
     */
    summary?: string | null;
    /**
     * @type null,string
     */
    storyline?: string | null;
    /**
     * @type null,integer, int64
     */
    updatedAt?: number | null;
    /**
     * @type null,integer, int64
     */
    firstReleaseDate?: number | null;
    /**
     * @type null,string
     */
    gameType?: string | null;
    /**
     * @type null,string
     */
    gameStatus?: string | null;
    /**
     * @type null,number, double
     */
    aggregatedRating?: number | null;
    /**
     * @type null,integer, int32
     */
    aggregatedRatingCount?: number | null;
    /**
     * @type null,number, double
     */
    rating?: number | null;
    /**
     * @type null,integer, int32
     */
    ratingCount?: number | null;
    /**
     * @type null,number, double
     */
    totalRating?: number | null;
    /**
     * @type null,integer, int32
     */
    totalRatingCount?: number | null;
    /**
     * @type null,string
     */
    coverUrl?: string | null;
    /**
     * @type null,array
     */
    themes?: string[] | null;
    /**
     * @type null,array
     */
    genres?: string[] | null;
    /**
     * @type null,array
     */
    gameModes?: string[] | null;
    /**
     * @type null,array
     */
    platforms?: string[] | null;
    /**
     * @type null,array
     */
    gameEngines?: string[] | null;
    /**
     * @type null,array
     */
    playerPerspectives?: string[] | null;
    /**
     * @type null,array
     */
    publishers?: string[] | null;
    /**
     * @type null,array
     */
    developers?: string[] | null;
    /**
     * @type null,array
     */
    multiplayerModes?: string[] | null;
    /**
     * @type null,integer, int64
     */
    totalVisits?: number | null;
    /**
     * @type null,number, double
     */
    igdbVisits?: number | null;
    /**
     * @type null,number, double
     */
    igdbWantToPlay?: number | null;
    /**
     * @type null,number, double
     */
    igdbPlaying?: number | null;
    /**
     * @type null,number, double
     */
    igdbPlayed?: number | null;
    /**
     * @type null,number, double
     */
    steam24hrPeakPlayers?: number | null;
    /**
     * @type null,number, double
     */
    steamPositiveReviews?: number | null;
    /**
     * @type null,number, double
     */
    steamNegativeReviews?: number | null;
    /**
     * @type null,number, double
     */
    steamTotalReviews?: number | null;
    /**
     * @type null,number, double
     */
    steamGlobalTopSellers?: number | null;
    /**
     * @type null,number, double
     */
    steamMostWishlistedUpcoming?: number | null;
    /**
     * @type null,number, double
     */
    twitch24hrHoursWatched?: number | null;
    /**
     * @type null,integer, int32
     */
    releaseYear?: number | null;
};